Refactor profile to use MVP
