+++ app/src/main/java/org/wit/hillfort/views/profile/ProfileView.kt
@@ -1,4 +1,4 @@
-package org.wit.hillfort.activities
+package org.wit.hillfort.views.profile
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,22 +6,28 @@ import com.squareup.picasso.Picasso
 import kotlinx.android.synthetic.main.activity_base.*
 import kotlinx.android.synthetic.main.activity_profile.*
 import org.wit.hillfort.R
+import org.wit.hillfort.activities.BaseActivity
 import org.wit.hillfort.helpers.CircleTransform
-import org.wit.hillfort.helpers.showImagePicker
 import org.wit.hillfort.models.user.UserModel
 
 
-class ProfileActivity : BaseActivity() {
-  var user = UserModel()
-  private val IMAGE_REQUEST = 1
+class ProfileView : BaseActivity() {
+  lateinit var presenter: ProfilePresenter
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     val contentView = layoutInflater.inflate(R.layout.activity_profile, null, false)
     drawer_layout.addView(contentView, 0)
 
-    // Get logged in user
-    user = app.currentUser!!
+    presenter = ProfilePresenter(this)
+
+    // Start image picker for profile image
+    iconLayout.setOnClickListener {
+      presenter.doSelectImage()
+    }
+  }
+
+  fun showProfile(user: UserModel) {
     profileName.text = "${user.name} ${user.surname}"
 
     // Load profile image
@@ -33,25 +39,13 @@ class ProfileActivity : BaseActivity() {
     // Load profile stats
     totalHillforts.text = "Total hillforts: ${app.hillforts.findAll().size}"
     totalVisited.text = "Total visited hillforts: ${app.hillforts.findVisited().size}"
-
-    // Start image picker for profile image
-    iconLayout.setOnClickListener {
-      showImagePicker(this, IMAGE_REQUEST)
-    }
   }
 
   // Recover the profile image when the image picker finishes
   override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
     super.onActivityResult(requestCode, resultCode, data)
-    when (requestCode) {
-      IMAGE_REQUEST -> {
-        if (data != null) {
-          user.profileImage = data.data.toString()
-          Picasso.get().load(user.profileImage)
-              .transform(CircleTransform()).into(profileIcon)
-          app.users.update(user)
-        }
-      }
+    if (data != null) {
+      presenter.doActivityResult(requestCode, resultCode, data)
     }
   }
 }
